Recalibrate distance and add questnav bootstrap

+++ src/main/java/binarycrows/robot/Utils/Gamepad/GenericGamepad.java
@@ -72,7 +72,7 @@ public class GenericGamepad {
         LogIOInputs.logToStateTable(value, "Gamepad/" + gamepad.getPort() + "/Axis/" + axis);
         return value;
     }
-    public double getPOV(int pov) {
+    public int getPOV(int pov) {
         return this.gamepad.getPOV(pov);
     }
